Revert "extract experimentation logic to a plugin"
This reverts commit 3c4ffe06a95a9914a49336ff03cef212f9d27143.
