+++ scripts/experimentation/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Adobe. All rights reserved.
+ * Copyright 2022 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -15,7 +15,7 @@ import {
   sampleRUM,
   toCamelCase,
   toClassName,
-} from '../../scripts/lib-franklin.js';
+} from '../lib-franklin.js';
 
 export const DEFAULT_OPTIONS = {
   root: '/experiments',
