+++ tools/preview/preview.css
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Adobe. All rights reserved.
+ * Copyright 2022 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -31,7 +31,7 @@
 }
 
 .hlx-badge {
-  border-radius: 32px;
+  border-radius: 32px; 
   background-color: #888;
   color: #eee;
   padding: 16px 32px;
@@ -104,7 +104,7 @@
 .hlx-popup-header {
   display: grid;
   grid-template:
-    "label actions"
+    "label actions" 
     "description actions"
     / 1fr min-content;
   background-color: #222;
@@ -149,7 +149,7 @@
   height: 0;
   position: absolute;
   border-left: 15px solid transparent;
-  border-right: 15px solid transparent;
+  border-right: 15px solid transparent;    
   border-top: 15px solid #444;
   bottom: -15px;
   right: 30px;
@@ -172,7 +172,7 @@
 .hlx-popup-item {
   display: grid;
   grid-template:
-    "label actions"
+    "label actions" 
     "description actions"
     / 1fr min-content;
   margin: 16px;
@@ -206,5 +206,5 @@
   .hlx-preview-overlay {
     flex-flow: row-reverse wrap-reverse;
     justify-content: flex-start;
-  }
+  }  
 }
